+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Change Log
 3.8.4 (unreleased)
 --------------------------
 
+### Improvements
+
+- Insert soft tabs to all insertion points when typing the tab key.
+
+
+
 ### Fixes
 
 - Fix an issue on macOS 10.13-14 where the application became instabil with some specific actions when the cursor locates the end of the document.